[refactor] remove patch_fxture, create TestSchema/Exception fixtures
